Calculate length correctly with ANSI escape codes
I was looking to do some more advanced string formatting using
[chalk](https://github.com/sindresorhus/chalk) in my grids/tables and
found cell length was being incorrectly calculated when you passed a
string that for example had been colorized, bolded and underlined.

This uses the strip-ansi package (which is powering the equivalent
functionality in Chalk) to calculate the length of the string to be
printed after first excluding any ansi escape sequences in the string.